Safari-611.1.10/LayoutTests/http/tests/websocket/tests/hybi/contentextensions/display-none-expected.txt

CONSOLE MESSAGE: onopen
CONSOLE MESSAGE: onmessage reply sent
CONSOLE MESSAGE: onclose
This should not be hidden.
